Thread configurable trustroot resync period to bundle trustroot func (#171)
* move trustroot resync period configration to different package


* add license


* comment


* rename files


---------
