+++ pkg/reconciler/trustroot/controller_test.go
@@ -16,10 +16,8 @@ package trustroot
 
 import (
 	"testing"
-	"time"
 
 	"knative.dev/pkg/configmap"
-	"knative.dev/pkg/controller"
 	rtesting "knative.dev/pkg/reconciler/testing"
 
 	// Fake injection informers
@@ -37,23 +35,5 @@ func TestNew(t *testing.T) {
 
 	if c == nil {
 		t.Fatal("Expected NewController to return a non-nil value")
-	}
-}
-
-func TestContextDuration(t *testing.T) {
-	ctx, _ := rtesting.SetupFakeContext(t)
-
-	expected := controller.DefaultResyncPeriod
-	actual := FromContextOrDefaults(ctx)
-	if expected != actual {
-		t.Fatal("Expected the context to store the value and be retrievable")
-	}
-
-	expected = time.Hour
-	ctx = ToContext(ctx, expected)
-	actual = FromContextOrDefaults(ctx)
-
-	if expected != actual {
-		t.Fatal("Expected the context to store the value and be retrievable")
 	}
 }
