+++ pkg/tuf/context_test.go
@@ -1,0 +1,42 @@
+//
+// Copyright 2024 The Sigstore Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package tuf
+
+import (
+	"testing"
+	"time"
+
+	"knative.dev/pkg/controller"
+	rtesting "knative.dev/pkg/reconciler/testing"
+)
+
+func TestContextDuration(t *testing.T) {
+	ctx, _ := rtesting.SetupFakeContext(t)
+
+	expected := controller.DefaultResyncPeriod
+	actual := FromContextOrDefaults(ctx)
+	if expected != actual {
+		t.Fatal("Expected the context to store the value and be retrievable")
+	}
+
+	expected = time.Hour
+	ctx = ToContext(ctx, expected)
+	actual = FromContextOrDefaults(ctx)
+
+	if expected != actual {
+		t.Fatal("Expected the context to store the value and be retrievable")
+	}
+}
